update: reformat packages
