--- domain/vod/src/main/java/model/Season.kt
@@ -1,8 +1,0 @@
-package model
-
-data class Season(
-    val id: String,
-    val title: String,
-    val number: Int,
-    val episodes: List<Episode>
-)
