+++ domain/vod/src/main/java/repository/ActorRepository.kt
@@ -1,7 +1,8 @@
 package repository
 
 import model.Actor
+import model.Language
 
 interface ActorRepository {
-    suspend fun searchActors(query: String, language: String): List<Actor>
+    suspend fun searchActors(query: String, language: Language): List<Actor>
 }
