+++ domain/vod/src/main/java/repository/TvSeriesRepository.kt
@@ -1,8 +1,9 @@
 package repository
 
+import model.Language
 import model.MediaFilters
 import model.TvSeries
 
 interface TvSeriesRepository {
-    suspend fun searchTvSeries(query: String, filters: MediaFilters?, language: String): List<TvSeries>
+    suspend fun searchTvSeries(query: String, filters: MediaFilters?, language: Language): List<TvSeries>
 }
