+++ domain/vod/src/main/java/usecase/SearchActorsUseCase.kt
@@ -3,6 +3,7 @@ import kotlinx.datetime.Clock
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.toLocalDateTime
 import model.Actor
+import model.Language
 import model.SearchCategory
 import model.SearchHistoryItem
 import repository.ActorRepository
@@ -13,7 +14,7 @@ class SearchActorsUseCase(
     private val actorRepo: ActorRepository,
     private val historyRepo: SearchHistoryRepository
 ) {
-    suspend fun execute(query: String, language: String): List<Actor> {
+    suspend fun execute(query: String, language: Language): List<Actor> {
         historyRepo.addSearchHistoryItem(
             SearchHistoryItem(
                 query,
