+++ domain/vod/src/main/java/usecase/SearchMoviesUseCase.kt
@@ -3,6 +3,7 @@ package usecase
 import kotlinx.datetime.Clock
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.toLocalDateTime
+import model.Language
 import model.MediaFilters
 import model.Movie
 import model.SearchCategory
@@ -14,7 +15,7 @@ class SearchMoviesUseCase(
     private val movieRepo: MovieRepository,
     private val historyRepo: SearchHistoryRepository
 ) {
-    suspend fun execute(query: String, filters: MediaFilters?, language: String): List<Movie> {
+    suspend fun execute(query: String, filters: MediaFilters?, language: Language): List<Movie> {
         historyRepo.addSearchHistoryItem(
             SearchHistoryItem(
                 query,
