+++ domain/vod/src/main/java/usecase/SearchTvSeriesUseCase.kt
@@ -2,6 +2,7 @@ package usecase
 import kotlinx.datetime.Clock
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.toLocalDateTime
+import model.Language
 import model.MediaFilters
 import model.SearchCategory
 import model.SearchHistoryItem
@@ -14,7 +15,7 @@ class SearchTvSeriesUseCase(
     private val tvRepo: TvSeriesRepository,
     private val historyRepo: SearchHistoryRepository
 ) {
-    suspend fun execute(query: String, filters: MediaFilters?, language: String): List<TvSeries> {
+    suspend fun execute(query: String, filters: MediaFilters?, language: Language): List<TvSeries> {
         historyRepo.addSearchHistoryItem(
             SearchHistoryItem(
                 query,
